Priority is called on both send/rec
In the case we get a message for something we don't care about anymore, we filter it out.  This makes it so when we switch scenes there's a small time where the sender hasn't realized yet (because lag) and still sends us stuff, eventhough locally we're in a different scene, so we filter out those message when rec'd and ignore them.